Update aiohttp_fetch.py
The person who runs this example and closes his application will have the RuntimeError exception that comes from https://github.com/CabbageDevelopment/qasync/blob/6c32bc9294c9ca93e42ba8175b9bf49f5254a0de/qasync/__init__.py#L374 this is caused by "awkreate_future. () "this stays alive when QApplication is closed, therefore the solution is to listen to the closing event of QApplication and cancel this future.

A permanent solution would be to replace the raise of the aforementioned line and replace it with future.cancel () and a return.