Only init already_running QEventLoops as running
When a QEventLoop is not already running, it should not set itself as
the running loop on construction, and should do so exclusively when
actually ran, i.e. in run_forever. Already running QEventLoops OTOH
should always set themselves as the running loop on construction.

This renders the set_running_loop constructor parameter obsolete. It is
kept for compatibility.