_IocpProactor: Acquire lock in previously missed methods
For the same reason that recv(), send(), accept() and connect()
need to acquire the lock, all the other operations need to as well.

This also shows a design limitation in this approach w.r.t.
maintainability (silent failure when new operations are added);
it would be preferable to either reproduce the windows_events
implementation in its entirety, or use proxying via composition
to ensure methods not made thread-safe cannot be invoked.

Finally, the _poll implementation seems to have further issues:
it is not clear to me that the _stopped_serving handling is
thread-safe,  _unregistered is not handled, and it seems
preferable not to hold the lock when calling
GetQueuedCompletionStatus.